requirements.txt
dot2tex==2.11.3
graphviz==0.20.1
hy==0.24.0
matplotlib==3.4.3
numpy==1.22.3
pandas==1.1.5
Pillow==9.2.0
scikit_learn==1.1.2
tensorflow==2.9.1
tqdm==4.64.0
pyyaml==5.3.1
